Bug Report 🐛
A bug is a problem causing a program/website to crash or produce invalid output. A bug report is a detailed report that details what is incorrect and needs to be fixed.  The report includes a request and/or specifics for how to solve each issue, as well as a list of causes or noticed faults to point out exactly what is perceived as incorrect.